+++ easy-extension/functions/tests/setup.ts
@@ -1,4 +1,5 @@
 import admin from "firebase-admin";
+import { UserModel } from "../src/models/user.model";
 // import serviceAccount from "../service-account.json";
 
 /**
@@ -9,6 +10,9 @@ import admin from "firebase-admin";
 export function initFirebaseAdminSDK() {
     if (admin.apps.length === 0) {
         admin.initializeApp(
+            {
+                databaseURL: 'https://withcenter-test-2-default-rtdb.asia-southeast1.firebasedatabase.app/',
+            }
             // {
             // credential: admin.credential.cert(serviceAccount as admin.ServiceAccount),
             // databaseURL:
@@ -24,7 +28,8 @@ export function initFirebaseAdminSDK() {
  * @param options optoins
  * @returns user
  */
-export async function createTestUser(options?: { email?: string | null; password?: string | null, phoneNumber?: string | null }) {
+export async function createTestUser(options?: { email?: string | null; password?: string | null, phoneNumber?: string | null, displayName?: string | null })
+    : Promise<admin.auth.UserRecord> {
     if (!options) options = {};
 
     // generate a random email address if one is not provided
@@ -43,9 +48,33 @@ export async function createTestUser(options?: { email?: string | null; password
 
     const user = await admin
         .auth()
-        .createUser({ email: options.email, password, phoneNumber: options.phoneNumber });
+        .createUser({
+            email: options.email, password, phoneNumber: options.phoneNumber,
+            displayName: options.displayName,
+        });
 
     //
     return user;
 }
 
+
+
+/**
+ * Create a user document under /users collection in Firestore and return the user document as in JSON.
+ */
+export async function createTestUserDocument(options?: { email?: string | null; password?: string | null, phoneNumber?: string | null, displayName?: string | null })
+    : Promise<Record<string, any>> {
+    const user = await createTestUser(options);
+    const data = {
+        uid: user.uid,
+        email: user.email ?? '',
+        phoneNumber: user.phoneNumber ?? '',
+        displayName: user.displayName ?? '',
+        photoUrl: user.photoURL ?? '',
+        createdAt: admin.firestore.FieldValue.serverTimestamp(),
+    };
+
+    await admin.firestore().collection("users").doc(user.uid).set(data);
+
+    return await UserModel.getDocument(user.uid) as Record<string, any>;
+}