+++ lib/src/model/post.dart
@@ -1,7 +1,6 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:fireflutter/fireflutter.dart';
-import 'package:flutter/material.dart';
 
 class Post with FirebaseHelper {
   final String id;
@@ -64,12 +63,14 @@ class Post with FirebaseHelper {
       'updatedAt': FieldValue.serverTimestamp(),
       'uid': myUid,
     };
-    // TODO review
-    debugPrint("weh t");
     final postId = PostService.instance.postCol.doc().id;
-    await PostService.instance.categoryCol.doc(postId).set(postData);
+    await PostService.instance.postCol.doc(postId).set(postData);
     postData['createdAt'] = Timestamp.now();
     postData['updatedAt'] = Timestamp.now();
     return Post.fromMap(map: postData, id: postId);
   }
+
+  @override
+  String toString() =>
+      'Post(id: $id, categoryId: $categoryId, title: $title, content: $content, uid: $uid, files: $files, createdAt: $createdAt, updatedAt: $updatedAt, likes: $likes, deleted: $deleted)';
 }