+++ lib/src/widget/category/category_list.dialog.dart
@@ -3,14 +3,19 @@ import 'package:flutter/material.dart';
 
 /// Full screen dialog to show list of categories
 class CategoryListDialog extends StatelessWidget {
-  const CategoryListDialog({super.key});
+  const CategoryListDialog({
+    super.key,
+    this.onTapCategory,
+  });
+
+  final Function(Category category)? onTapCategory;
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
         backgroundColor: Theme.of(context).colorScheme.inversePrimary,
-        title: const Text('Categories'),
+        title: const Text("Categories"),
         actions: [
           IconButton(
             icon: const Icon(Icons.add),
@@ -26,7 +31,15 @@ class CategoryListDialog extends StatelessWidget {
           ),
         ],
       ),
-      body: const CategoryListView(),
+      body: CategoryListView(
+        onTap: (category) {
+          if (onTapCategory != null) {
+            onTapCategory?.call(category);
+          } else {
+            PostService.instance.showPostListDialog(context, category);
+          }
+        },
+      ),
     );
   }
 }